实现完整的判题逻辑和结果存储及获取的逻辑

+++ domain/src/main/java/org/caesar/domain/executor/request/ExecuteCodeRequest.java
@@ -30,7 +30,7 @@ public class ExecuteCodeRequest {
     private long questionId;
 
     // 提交ID
-    private long submitId;
+    private int submitId;
 
     // 用户代码
     private String code;
